Can select active workspace.
Workspace context will first try to get the workspace from the localstorage. Then it will make a request to get all the workspaces available to the user. It will then set the workspace if it can find it. Clicking on the workspace tile on the left hand side will select the workspace. Active workspace is shown in the TopBar.