Added block store and fixed loop. (#12)
* Added block store and fixed loop.

Fixed issue where objects in the object store were not being looped over fully. Added the block store tab which simply displays the data on a block store object.

* Added reusable table component.

* Addressed review comments.

Made the table a reusable component. Made the block stores into a table using this new component. Added a link to /apphub as Alex H's suggestion for now. Added the link to the block store description in the header,.

* Added guards incase of null data.

* Addressed review comments.

* Moved styles to table component.