Now use real data from api.
Local data as per usual is stored in placeholder.ts. This is in the correct form that you would expect from the api call . Added a chart using chart.js, this makes it so that we can simply show the data without having to write our own interactive stacked bar chart. Took insiration from AWS and how they do their billing in the UI. Currently only show the current month and previous month worth of billing. It can be extended to show previous months / years but this would require a fair amount of work.