1094 account usage information (#20)
* Added warnings when no billing account or workspace.

Disabled the ability to add a workspace when no billing account. Fixed credentials breaking the site when no workspace selected. User cannot generate an API token when no workspace is selected. Unrelated but fixed dropdown styling.

* Linting.

* Added chart and dummy data for invoices.

* Linting.

* linting again.

* Now use real data from api.

Local data as per usual is stored in placeholder.ts. This is in the correct form that you would expect from the api call . Added a chart using chart.js, this makes it so that we can simply show the data without having to write our own interactive stacked bar chart. Took insiration from AWS and how they do their billing in the UI. Currently only show the current month and previous month worth of billing. It can be extended to show previous months / years but this would require a fair amount of work.

* Added comment for sku reading.

* Added comment for long useEffect.

* Bug fix.

* Y axis + removed undefined when no total.

---------

Co-authored-by: Ubuntu <ubuntu@ip-10-2-255-14.eu-west-2.compute.internal>