1271 linked accounts contract validation (#27)
* Airbus now validates key before submitting.

Once validated will show dropdowns for legacy and pneo.

* Fixed style linting.

* Uses validation response data.

Will get response data and use this to construct the options inside of the select elements. If there is only one option the dropdown will take this value and be disabled. If the account later being linked is airbus then it adds the contract data onto the body.

* Linting.

* removed getAccounts call that reset state too early.

* Will show correct button after validation

* Using correct key when linking account after validation.

* Code tidy.

* Reset error before validating.

* Addressed review comments.

* linting.