Can account for usage.
When we get an SKU, we look at its unit. If there is no unit associated to the SKU then we just display the usage. If the SKU has a unit it will look for a conversion method to convert this quantity + unit into pounds, if no method is found then it will default back to just showing usage.