1241 account data (#40)
* Added invoices in navigation.

* Can account for usage.

When we get an SKU, we look at its unit. If there is no unit associated to the SKU then we just display the usage. If the SKU has a unit it will look for a conversion method to convert this quantity + unit into pounds, if no method is found then it will default back to just showing usage.

* Will only use costs if has a unit and conversion method.

* Linting.

* eslint redundant line.

* Added .dockerignore file.

* Removed .dockerignore as not required.

* Will only make single response for each sku type.

* Fixed  bugs.

* Now show current month usage only.

* Use current month.

* Add warnings and info to invoices to help user.

* linting.

* Added catalog name to collection file name using _src/ as the string separator.

* Fix comparison and just use usage.

* Added use of costs.

* Does not show labels if data is zero. Shows current monthly usage and costs.

* Added table component to invoices.

* Ignore redundant.

* Will fetch all skus.

* Added price and limit to 3 sig figs.